requirements.txt
django-ckeditor-5==0.1.6 
Pillow==9.2.0 